Call the OpenWeather API to get weather information

+++ forecast/forecast.go
@@ -1,6 +1,6 @@
 package forecast
 
 type Forecast struct {
-	ZipCode string
-	CurrentTemperature int
+	ZipCode            string
+	CurrentTemperature float32
 }
